Integrate New Time Filter
